c: environment ops in env.c
